// frontend/src/app/api/model/paymentMethod.ts
// File generated using spec-to-code.sh, DO NOT EDIT MANUALLY.

/**
 * ADH6 API
 *
 * Contact: equipe@minet.net
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PaymentMethod { 
    /**
     * The unique identifier of this payment method
     */
    readonly id?: number;
    /**
     * The static fee for this payment method in euros
     */
    staticFee?: number | null;
    /**
     * The dynamic fee for this payment method as a percentage
     */
    dynamicFee?: number | null;
    /**
     * The friendly name of this payment method
     */
    name: string;
}